requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
pyautogui>=0.9.54
numpy>=1.24.0
Pillow>=10.0.0
tk>=0.1.0